admin

+++ database/migrations/2025_02_20_062903_create_product_variants_table.php
@@ -15,8 +15,8 @@ return new class extends Migration {
             $table->foreignId('product_id')->constrained()->onDelete('cascade');
             $table->string('color');
             $table->string('storage');
-            $table->decimal('origin_price', 15, 2)->nullable();
-            $table->decimal('price', 15, 2)->nullable();
+            $table->integer('origin_price')->nullable();
+            $table->integer('price')->nullable();
             $table->integer('stock')->nullable();
             $table->enum('status', ['active', 'inactive'])->default('active');
             $table->timestamps();
